Remove the EditorUI
